Topic validation & autologout finished
